made requested changes to service modules
